requirements.txt

Flask==3.0.0
cs50==9.2.5
werkzeug==3.0.0
psycopg2-binary==2.9.9


